Separate API and update types
