--- src/volkovlabs-balenasupervisor-datasource/api/models/index.ts
@@ -1,4 +1,0 @@
-export * from './device';
-export * from './log-entry';
-export * from './state-status';
-export * from './target-state';
