+++ src/volkovlabs-balenasupervisor-datasource/types/device.ts
