+++ src/volkovlabs-balenasupervisor-datasource/types/index.ts
@@ -1,2 +1,6 @@
 export * from './datasource';
+export * from './device';
+export * from './log-entry';
 export * from './query';
+export * from './state-status';
+export * from './target-state';
