+++ src/volkovlabs-balenasupervisor-datasource/types/log-entry.ts
