+++ src/volkovlabs-balenasupervisor-datasource/types/state-status.ts
